Account for multiple ways of overwriting methods
In order to attempt to prevent as many possible ways of detecting cheating as possible, a banned_terms array was built to contain any keyword or identifier a user should not be able to enter. If their code contains one of those tokens, an exception is raised. If other methods of cheating are discovered, they can be added to the banned_terms array.